refactor: Make flow control syncing default and drop the no flow parts
