+++ dash-spv/src/network/tests.rs
@@ -27,23 +27,11 @@ mod peer_network_manager_tests {
             max_peers: 3,
             enable_persistence: false,
             log_level: "info".to_string(),
-            enable_filter_flow_control: true,
             filter_request_delay_ms: 0,
             max_concurrent_filter_requests: 50,
             max_concurrent_cfheaders_requests_parallel: 50,
-            enable_cfheaders_flow_control: true,
             cfheaders_request_timeout_secs: 30,
             max_cfheaders_retries: 3,
-            enable_cfheader_gap_restart: true,
-            cfheader_gap_check_interval_secs: 15,
-            cfheader_gap_restart_cooldown_secs: 30,
-            max_cfheader_gap_restart_attempts: 5,
-            enable_filter_gap_restart: true,
-            filter_gap_check_interval_secs: 20,
-            min_filter_gap_size: 10,
-            filter_gap_restart_cooldown_secs: 30,
-            max_filter_gap_restart_attempts: 5,
-            max_filter_gap_sync_size: 50000,
             // Mempool fields
             enable_mempool_tracking: false,
             mempool_strategy: crate::client::config::MempoolStrategy::BloomFilter,
